+++ mainwindow.h
@@ -14,8 +14,11 @@ class MainWindow : public QMainWindow
 public:
     MainWindow(QWidget *parent = nullptr);
     ~MainWindow();
-
+public slots:
+    void onClickMe();
 private:
     Ui::MainWindow *ui;
+
+    int lable_number=0;
 };
 #endif // MAINWINDOW_H
